+++ src/main/java/umc/cockple/demo/domain/exercise/controller/api/ExerciseMyApi.java
@@ -8,10 +8,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import umc.cockple.demo.domain.exercise.dto.MyExerciseCalendarDTO;
-import umc.cockple.demo.domain.exercise.dto.MyExerciseListDTO;
-import umc.cockple.demo.domain.exercise.dto.MyPartyExerciseCalendarDTO;
-import umc.cockple.demo.domain.exercise.dto.MyPartyExerciseDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyExerciseCalendarDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyExerciseListDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyPartyExerciseCalendarDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyPartyExerciseDTO;
 import umc.cockple.demo.domain.exercise.enums.MyExerciseFilterType;
 import umc.cockple.demo.domain.exercise.enums.MyExerciseOrderType;
 import umc.cockple.demo.domain.exercise.enums.MyPartyExerciseOrderType;
