+++ src/main/java/umc/cockple/demo/domain/exercise/dto/my/MyExerciseListDTO.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.exercise.dto;
+package umc.cockple.demo.domain.exercise.dto.my;
 
 import lombok.Builder;
