+++ src/test/java/umc/cockple/demo/domain/exercise/service/ExerciseLifecycleQueryServiceTest.java
@@ -23,10 +23,10 @@ import umc.cockple.demo.domain.exercise.dto.lifecycle.ExerciseDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.lifecycle.ExerciseEditDetailDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseMapBuildingsDTO;
 import umc.cockple.demo.domain.exercise.dto.guest.ExerciseMyGuestListDTO;
-import umc.cockple.demo.domain.exercise.dto.MyExerciseCalendarDTO;
-import umc.cockple.demo.domain.exercise.dto.MyExerciseListDTO;
-import umc.cockple.demo.domain.exercise.dto.MyPartyExerciseCalendarDTO;
-import umc.cockple.demo.domain.exercise.dto.MyPartyExerciseDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyExerciseCalendarDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyExerciseListDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyPartyExerciseCalendarDTO;
+import umc.cockple.demo.domain.exercise.dto.my.MyPartyExerciseDTO;
 import umc.cockple.demo.domain.exercise.dto.PartyExerciseCalendarDTO;
 import umc.cockple.demo.domain.exercise.enums.MyExerciseFilterType;
 import umc.cockple.demo.domain.exercise.enums.MyExerciseOrderType;
